requirements.txt
StructPy~=0.1
numpy~=1.22.3
PyYAML~=6.0
pytest~=7.1.1
setuptools~=60.2.0